extract file and add specific dedup functions
